Move `dedent` to its own file
