+++ src/conversion/importExecution/importExecutionConverter.spec.ts
@@ -3,7 +3,7 @@ import { readFileSync } from "fs";
 import { stubLogging } from "../../../test/util";
 import { initOptions } from "../../context";
 import { InternalOptions } from "../../types/plugin";
-import { dedent } from "../../types/util";
+import { dedent } from "../../util/dedent";
 import { TestIssueData } from "./importExecutionConverter";
 import { ImportExecutionConverterCloud } from "./importExecutionConverterCloud";
 import { ImportExecutionConverterServer } from "./importExecutionConverterServer";
