+++ src/hooks.spec.ts
@@ -9,7 +9,7 @@ import { XrayClientServer } from "./client/xray/xrayClientServer";
 import { initOptions } from "./context";
 import { afterRunHook, beforeRunHook, synchronizeFile } from "./hooks";
 import { ClientCombination, InternalOptions } from "./types/plugin";
-import { dedent } from "./types/util";
+import { dedent } from "./util/dedent";
 
 // Enable promise assertions.
 chai.use(chaiAsPromised);
